refactor: 인스턴스 목록 컴포넌트 구조 개선
