+++ src/components/instance/list/InstanceTable.vue
@@ -1,0 +1,261 @@
+<template>
+  <div>
+    <!-- 에러 메시지 -->
+    <div v-if="error" class="error-message">
+      {{ error }}
+    </div>
+
+    <!-- 로딩 상태 -->
+    <div v-if="isLoading" class="loading-state">데이터를 불러오는 중...</div>
+
+    <!-- 인스턴스 테이블 -->
+    <div v-else class="table-container">
+      <table class="instance-table">
+        <thead>
+          <tr>
+            <!-- 체크박스 전체 선택 -->
+            <th class="checkbox-column">
+              <input
+                type="checkbox"
+                @change="handleSelectAllInstances"
+                :checked="isAllSelected"
+                :disabled="!instances.length"
+              />
+            </th>
+            <th>인스턴스</th>
+            <th>전원상태</th>
+            <th>CPU</th>
+            <th>메모리</th>
+            <th>디스크</th>
+            <th>IP 주소</th>
+            <th>전원관리</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr v-if="!instances.length">
+            <td colspan="8" class="no-data">인스턴스가 없습니다.</td>
+          </tr>
+          <tr v-else v-for="instance in instances" :key="instance.id">
+            <td class="checkbox-column">
+              <input
+                type="checkbox"
+                :value="instance.id"
+                :checked="selectedIds.includes(instance.id)"
+                @change="handleSelectInstance(instance.id)"
+              />
+            </td>
+            <td>
+              <router-link
+                :to="{ name: 'instanceDetail', params: { id: instance.id } }"
+                class="instance-link"
+              >
+                {{ instance.name }}
+              </router-link>
+            </td>
+            <td>
+              <StatusBadge :status="instance.status" />
+            </td>
+            <td>{{ instance.cpu }}코어</td>
+            <td>{{ instance.memory }}GB</td>
+            <td>{{ instance.disk }}GB</td>
+            <td>{{ instance.privateIp }}</td>
+            <td>
+              <PowerToggle :is-running="instance.powerOn" @toggle="handlePowerToggle(instance)" />
+            </td>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  </div>
+</template>
+
+<script setup lang="ts">
+import { ref, computed } from 'vue'
+import type { Instance } from '@/mock/types/instance'
+import StatusBadge from './StatusBadge.vue'
+import PowerToggle from './PowerToggle.vue'
+
+interface Props {
+  instances: Instance[]
+  isLoading?: boolean
+  error?: string | null
+}
+
+const props = withDefaults(defineProps<Props>(), {
+  isLoading: false,
+  error: null,
+})
+
+const emit = defineEmits<{
+  (e: 'select', ids: string[]): void
+  (e: 'togglePower', instance: Instance): void
+}>()
+
+// 현재 선택된 모든 인스턴스 ID들
+const selectedIds = ref<string[]>([])
+
+// 모든 인스턴스가 선택되었는지 확인
+const isAllSelected = computed(() => {
+  return props.instances.length > 0 && selectedIds.value.length === props.instances.length
+})
+
+// 전체 인스턴스 선택/해제 처리
+const handleSelectAllInstances = (event: Event) => {
+  const target = event.target as HTMLInputElement
+  if (target.checked) {
+    selectedIds.value = props.instances.map((instance) => instance.id)
+  } else {
+    selectedIds.value = []
+  }
+
+  // InstanceListView에 이벤트 전달
+  emit('select', selectedIds.value)
+}
+
+// 체크박스 선택/해제 처리
+const handleSelectInstance = (id: string) => {
+  const index = selectedIds.value.indexOf(id)
+  if (index === -1) {
+    selectedIds.value.push(id) // 체크박스 선택
+  } else {
+    selectedIds.value.splice(index, 1) // 체크박스 해제
+  }
+  emit('select', selectedIds.value)
+}
+
+// 전원 토글 처리
+const handlePowerToggle = (instance: Instance) => {
+  emit('togglePower', instance)
+}
+</script>
+
+<style scoped>
+/* 테이블 스타일 */
+.table-container {
+  border: 1px solid #d9d9d9;
+  border-radius: 8px;
+  overflow: hidden;
+  background-color: #ffffff;
+}
+
+.instance-table {
+  width: 100%;
+  border-collapse: collapse;
+  font-size: 14px;
+}
+
+.instance-table th {
+  background-color: #fafafa;
+  padding: 12px 16px;
+  text-align: center;
+  font-weight: 600;
+  color: #262626;
+  border-bottom: 1px solid #d9d9d9;
+  border-right: 1px solid #d9d9d9;
+}
+
+.instance-table th:last-child {
+  border-right: none;
+}
+
+.instance-table td {
+  padding: 12px 16px;
+  text-align: center;
+  border-bottom: 1px solid #f0f0f0;
+  border-right: 1px solid #f0f0f0;
+  color: #262626;
+}
+
+.instance-table td:last-child {
+  border-right: none;
+}
+
+.instance-table tbody tr:hover {
+  background-color: #f5f5f5;
+}
+
+.instance-table tbody tr:last-child td {
+  border-bottom: none;
+}
+
+.checkbox-column {
+  width: 60px;
+}
+
+.instance-link {
+  color: #1890ff;
+  text-decoration: none;
+}
+
+.instance-link:hover {
+  text-decoration: underline;
+}
+
+.error-message {
+  color: #ff4d4f;
+  padding: 16px;
+  text-align: center;
+  background-color: #fff2f0;
+  border: 1px solid #ffccc7;
+  border-radius: 8px;
+  margin-bottom: 16px;
+}
+
+.loading-state {
+  text-align: center;
+  padding: 24px;
+  color: #8c8c8c;
+}
+
+.no-data {
+  text-align: center;
+  color: #8c8c8c;
+  padding: 24px;
+}
+
+/* 체크박스 스타일 */
+input[type='checkbox'] {
+  width: 16px;
+  height: 16px;
+  cursor: pointer;
+}
+
+/* 반응형 스타일 */
+@media (max-width: 768px) {
+  .instance-table {
+    font-size: 12px;
+  }
+
+  .instance-table th,
+  .instance-table td {
+    padding: 8px 12px;
+  }
+
+  /* 모바일에서는 CPU, 메모리, 디스크 열 숨김 */
+  .instance-table th:nth-child(4),
+  .instance-table td:nth-child(4),
+  .instance-table th:nth-child(5),
+  .instance-table td:nth-child(5),
+  .instance-table th:nth-child(6),
+  .instance-table td:nth-child(6) {
+    display: none;
+  }
+}
+
+@media (min-width: 769px) and (max-width: 1024px) {
+  .instance-table {
+    font-size: 13px;
+  }
+
+  .instance-table th,
+  .instance-table td {
+    padding: 10px 14px;
+  }
+
+  /* 태블릿에서는 디스크 열만 숨김 */
+  .instance-table th:nth-child(6),
+  .instance-table td:nth-child(6) {
+    display: none;
+  }
+}
+</style>
