+++ src/mock/data/instances.json
@@ -30,21 +30,6 @@
       "publicIp": null
     },
     {
-      "id": "inst-005",
-      "name": "test-environment",
-      "status": "RUNNING",
-      "flavor": "C2M4D40",
-      "image": "ubuntu-2204-kube-v1.30.14",
-      "network": "public-network",
-      "cpu": 2,
-      "memory": 4,
-      "disk": 40,
-      "privateIp": "172.30.110.86",
-      "powerOn": true,
-      "createdAt": "2024-02-05T16:30:00Z",
-      "publicIp": "203.0.113.48"
-    },
-    {
       "id": "inst-006",
       "name": "monitoring-server",
       "status": "RUNNING",
@@ -58,20 +43,6 @@
       "powerOn": true,
       "createdAt": "2024-02-10T11:20:00Z",
       "publicIp": "203.0.113.94"
-    },
-    {
-      "id": "inst-1752026407125",
-      "name": "test",
-      "image": "ubuntu-2204-kube-v1.30.9",
-      "flavor": "C4M8D80",
-      "network": "private-network",
-      "cpu": 4,
-      "memory": 8,
-      "disk": 80,
-      "status": "SHUTDOWN",
-      "powerOn": false,
-      "createdAt": "2025-07-09T02:00:07.124Z",
-      "privateIp": "172.30.110.169"
     },
     {
       "id": "5839",
@@ -93,21 +64,22 @@
       "name": "진짜 테스트",
       "image": "ubuntu-2204-kube-v1.29.7",
       "flavor": "C2M4D40",
-      "network": "private-network",
+      "network": "public-network",
       "cpu": 2,
       "memory": 4,
       "disk": 40,
       "status": "RUNNING",
       "powerOn": true,
       "createdAt": "2025-07-09T09:06:49.190Z",
-      "privateIp": "172.30.110.149"
+      "privateIp": "172.30.110.149",
+      "publicIp": "203.0.113.203"
     },
     {
       "id": "afdb",
       "name": "진짜 테스트2",
       "image": "ubuntu-2204-kube-v1.28.5",
       "flavor": "C2M4D40",
-      "network": "public-network",
+      "network": "private-network",
       "cpu": 2,
       "memory": 4,
       "disk": 40,
@@ -115,7 +87,7 @@
       "powerOn": true,
       "createdAt": "2025-07-09T09:07:59.660Z",
       "privateIp": "172.30.110.211",
-      "publicIp": "203.0.113.96"
+      "publicIp": null
     }
   ]
 }
